Crate secret_toolkit_storage

Re-exports

pub use append_store::AppendStore;

pub use append_store::AppendStoreMut;

pub use deque_store::DequeStore;

pub use deque_store::DequeStoreMut;

pub use typed_store::TypedStore;

pub use typed_store::TypedStoreMut;

Modules

append_store

An “append store” is a storage wrapper that guarantees constant-cost appending to and popping from a list of items in storage.

deque_store

This is a storage wrapper based on AppendStore called DequeStore. It guarantees constant-cost appending to and popping from a list of items in storage on both directions (front and back).

typed_store